fix(examples): make every FloatType alias actually switch precision
Six mathematics examples declared `type FloatType` and could not compile at
`Float106`. Each built at `f64`, so the suite was green and the README's claim
that one edit re-runs the example at a new precision was false for most of the
tree.

Four causes, none of which fails at f64:
  - a `const` holding the working type; a `const` takes a primitive literal
    only, so `const ALPHA: FloatType = 0.15` cannot hold a software scalar
  - an inherent `.sqrt()`, `.powi()`, `.abs()`, `.cos()` called without
    `deep_causality_algebra::Real` or `deep_causality_num::Float` in scope
  - an accumulator seeded with a bare literal, which infers `f64`
  - `SimplicialManifold<f64, FloatType>`, pinning the complex's coefficients
    regardless of the alias

Repaired: triple_hkt_stress_field, effect_diffusion_on_manifold,
differential_field, cubical_heat_diffusion, hkt_graph_convolution,
lattice_gauge_simulation. Coefficient leaks also fixed in
tensor_x_topology_laplacian and capstone_spinor_minkowski.
multifield_data_pipeline's `1e-6`/`1e-5` thresholds now derive from
`<FloatType as Real>::epsilon()`, so they move with the alias.

Numerical output at f64 is unchanged; triple_hkt_stress_field still reports
2.403e8 Pa at the two vertices with the largest x.

`scripts/check_precision.sh` (`make check_precision`) flips every declared
alias through Float106/f32/f64 and rebuilds, so the claim is enforced rather
than asserted. 14 of 14 aliased examples pass.

+++ examples/mathematics_examples/composable_multi_math/capstone_spinor_minkowski/main.rs
@@ -111,7 +111,7 @@ fn main() {
 
 /// Build the discretized timelike path manifold with per-edge rapidities stored in the
 /// data tensor. Vertex entries hold zero; edge entries hold the rapidity.
-fn build_path_manifold(rapidities: &[FloatType]) -> SimplicialManifold<f64, FloatType> {
+fn build_path_manifold(rapidities: &[FloatType]) -> SimplicialManifold<FloatType, FloatType> {
     assert_eq!(rapidities.len(), N_EDGES);
 
     let vertices: Vec<Simplex> = (0..N_VERTICES).map(|i| Simplex::new(vec![i])).collect();
@@ -141,7 +141,7 @@ fn build_path_manifold(rapidities: &[FloatType]) -> SimplicialManifold<f64, Floa
 /// comonadic cursor and using `extract`. This demonstrates the topology and
 /// tensor cooperating: the manifold provides the addressing, the tensor
 /// provides the storage.
-fn read_edge_rapidity(m: &SimplicialManifold<f64, FloatType>, e: usize) -> FloatType {
+fn read_edge_rapidity(m: &SimplicialManifold<FloatType, FloatType>, e: usize) -> FloatType {
     let cursor = N_VERTICES + e;
     let repositioned =
         Manifold::new(m.complex().clone(), m.data().clone(), cursor).expect("reposition");
@@ -176,7 +176,7 @@ fn boost_rotor(theta: FloatType) -> (CausalMultiVector<FloatType>, CausalMultiVe
 /// manifold, builds the rotor, and applies `psi -> B psi B~`.
 fn transport_across_edge(
     psi: CausalMultiVector<FloatType>,
-    manifold: &SimplicialManifold<f64, FloatType>,
+    manifold: &SimplicialManifold<FloatType, FloatType>,
     e: usize,
 ) -> Process<CausalMultiVector<FloatType>> {
     let theta = read_edge_rapidity(manifold, e);
